+++ src/com/yandex/app/service/InMemoryTaskManager.java
@@ -110,7 +110,7 @@ public class InMemoryTaskManager implements TaskManager {
 
     @Override
     public String deleteOneTaskByID(int idForDelete) {
-        inMemoryHistoryManager.removeOneElem(idForDelete);
+        inMemoryHistoryManager.remove(idForDelete);
         allTasks.remove(idForDelete);
         return printMessageAboutSuccessfulFinishingOperation();
     }
@@ -118,10 +118,10 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public String deleteOneEpicByID(int idForDelete) {
         for (int index : allEpics.get(idForDelete).getSubtasksIDs()) {
-            inMemoryHistoryManager.removeOneElem(index);
+            inMemoryHistoryManager.remove(index);
             allSubtasks.remove(index);
         }
-        inMemoryHistoryManager.removeOneElem(idForDelete);
+        inMemoryHistoryManager.remove(idForDelete);
         allEpics.remove(idForDelete);
         return printMessageAboutSuccessfulFinishingOperation();
     }
@@ -129,7 +129,7 @@ public class InMemoryTaskManager implements TaskManager {
     @Override
     public String deleteOneSubtaskskByID(int idForDelete) {
         int idOfEpic = allSubtasks.get(idForDelete).getIdOfSubtaskEpic();
-        inMemoryHistoryManager.removeOneElem(idForDelete);
+        inMemoryHistoryManager.remove(idForDelete);
         allEpics.get(idOfEpic).getSubtasksIDs().remove(Integer.valueOf(idForDelete));
         allSubtasks.remove(idForDelete);
         checkProgressStatusOfEpic(idOfEpic);
