--- src/com/yandex/app/service/model/Node.java
@@ -1,18 +1,0 @@
-package com.yandex.app.service.model;
-
-public class Node<T> {
-    public T data;
-    public Node<T> next;
-    public Node<T> prev;
-
-    public Node(Node<T> prev, T data, Node<T> next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-
-    @Override
-    public String toString() {
-        return data.toString();
-    }
-}
